Update documentation for alb lambda
